move banner to its own module
